requirements.txt
streamlit
google-cloud-bigquery
pandas-gbq
seatsio
pandas